Online Bookstore/PracticeWebsite/requirement.txt
flask
flask_sqlalchemy
flask_wtf
flask_login
email_validator
boto3